[chore] able to add a study log
